Assignment E2 Complete

+++ src/fr222cy_assign2/Exercise_2/Bicycle.java
@@ -4,7 +4,7 @@ public class Bicycle extends Vehicle{
 	private final int MAX_PASSENGERS = 1;
 	public Bicycle(){
 		this.setCost(40);
-		this.setSpace((float) 0.2);
+		this.setSpace((double) 0.2);
 		this.setNumOfPassengers(1);
 	}
 	@Override
